backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
sqlalchemy==2.0.37
alembic==1.15.0
python-dotenv==1.0.1
pydantic-settings==2.8.1
# psycopg2-binary==2.9.10  # uncomment for PostgreSQL
